+++ Elements/src/Geometry/CsgExtensions.cs
@@ -29,7 +29,6 @@ namespace Elements.Geometry
             out double[] vmax, out double[] vmin, out double[] nmin, out double[] nmax,
             out float[] cmin, out float[] cmax, out ushort imin, out ushort imax, out double[] uvmin, out double[] uvmax)
         {
-
             var tessellations = new Tess[csg.Polygons.Count];
 
             var fi = 0;
@@ -65,9 +64,8 @@ namespace Elements.Geometry
             nmin = new double[3] { double.MaxValue, double.MaxValue, double.MaxValue };
             nmax = new double[3] { double.MinValue, double.MinValue, double.MinValue };
 
-            // TODO: Set this properly when solids get UV coordinates.
-            uvmin = new double[2] { 0, 0 };
-            uvmax = new double[2] { 0, 0 };
+            uvmin = new double[2] { double.MaxValue, double.MaxValue };
+            uvmax = new double[2] { double.MinValue, double.MinValue };
 
             imax = ushort.MinValue;
             imin = ushort.MaxValue;
@@ -77,6 +75,8 @@ namespace Elements.Geometry
             var uvi = 0;
 
             var iCursor = 0;
+
+            (Vector3 U, Vector3 V) basis;
 
             for (var i = 0; i < tessellations.Length; i++)
             {
@@ -88,23 +88,33 @@ namespace Elements.Geometry
                 var a = tess.Vertices[tess.Elements[0]].Position.ToVector3();
                 var b = tess.Vertices[tess.Elements[1]].Position.ToVector3();
                 var c = tess.Vertices[tess.Elements[2]].Position.ToVector3();
-                var n = (b - a).Cross(c - a).Unitized();
+                var tmp = (b - a).Unitized();
+                var n = tmp.Cross(c - a).Unitized();
+
+                // Calculate the texture space basis vectors
+                // from the first triangle. This is acceptable
+                // for planar faces.
+                // TODO: Update this when we support non-planar faces.
+                // https://gamedev.stackexchange.com/questions/172352/finding-texture-coordinates-for-plane
+                basis = n.ComputeDefaultBasisVectors();
 
                 for (var j = 0; j < tess.Vertices.Length; j++)
                 {
                     var v = tess.Vertices[j];
+                    var p = v.Position.ToVector3();
 
-                    System.Buffer.BlockCopy(BitConverter.GetBytes((float)v.Position.X), 0, vertexBuffer, vi, floatSize);
-                    System.Buffer.BlockCopy(BitConverter.GetBytes((float)v.Position.Y), 0, vertexBuffer, vi + floatSize, floatSize);
-                    System.Buffer.BlockCopy(BitConverter.GetBytes((float)v.Position.Z), 0, vertexBuffer, vi + 2 * floatSize, floatSize);
+                    System.Buffer.BlockCopy(BitConverter.GetBytes((float)p.X), 0, vertexBuffer, vi, floatSize);
+                    System.Buffer.BlockCopy(BitConverter.GetBytes((float)p.Y), 0, vertexBuffer, vi + floatSize, floatSize);
+                    System.Buffer.BlockCopy(BitConverter.GetBytes((float)p.Z), 0, vertexBuffer, vi + 2 * floatSize, floatSize);
 
                     System.Buffer.BlockCopy(BitConverter.GetBytes((float)n.X), 0, normalBuffer, vi, floatSize);
                     System.Buffer.BlockCopy(BitConverter.GetBytes((float)n.Y), 0, normalBuffer, vi + floatSize, floatSize);
                     System.Buffer.BlockCopy(BitConverter.GetBytes((float)n.Z), 0, normalBuffer, vi + 2 * floatSize, floatSize);
 
-                    // TODO: Update Solids to use something other than UV = {0,0}.
-                    System.Buffer.BlockCopy(BitConverter.GetBytes(0f), 0, uvBuffer, uvi, floatSize);
-                    System.Buffer.BlockCopy(BitConverter.GetBytes(0f), 0, uvBuffer, uvi + floatSize, floatSize);
+                    var uu = basis.U.Dot(p);
+                    var vv = basis.V.Dot(p);
+                    System.Buffer.BlockCopy(BitConverter.GetBytes((float)uu), 0, uvBuffer, uvi, floatSize);
+                    System.Buffer.BlockCopy(BitConverter.GetBytes((float)vv), 0, uvBuffer, uvi + floatSize, floatSize);
 
                     uvi += 2 * floatSize;
                     vi += 3 * floatSize;
@@ -123,10 +133,10 @@ namespace Elements.Geometry
                     nmin[1] = Math.Min(nmin[1], n.Y);
                     nmin[2] = Math.Min(nmin[2], n.Z);
 
-                    // uvmax[0] = Math.Max(uvmax[0], 0);
-                    // uvmax[1] = Math.Max(uvmax[1], 0);
-                    // uvmin[0] = Math.Min(uvmin[0], 0);
-                    // uvmin[1] = Math.Min(uvmin[1], 0);
+                    uvmax[0] = Math.Max(uvmax[0], uu);
+                    uvmax[1] = Math.Max(uvmax[1], vv);
+                    uvmin[0] = Math.Min(uvmin[0], uu);
+                    uvmin[1] = Math.Min(uvmin[1], vv);
                 }
 
                 for (var k = 0; k < tess.Elements.Length; k++)
@@ -175,6 +185,9 @@ namespace Elements.Geometry
 
                 tess.Tessellate(WindingRule.Positive, LibTessDotNet.Double.ElementType.Polygons, 3);
 
+                Vector3 e1 = new Vector3();
+                Vector3 e2 = new Vector3();
+
                 var vertices = new List<Csg.Vertex>();
                 for (var i = 0; i < tess.ElementCount; i++)
                 {
@@ -204,23 +217,31 @@ namespace Elements.Geometry
                             cv = v;
                         }
                     }
+                    if (i == 0)
+                    {
+                        var n = f.Plane().Normal;
+                        e1 = n.Cross(n.IsParallelTo(Vector3.XAxis) ? Vector3.YAxis : Vector3.XAxis).Unitized();
+                        e2 = n.Cross(e1).Unitized();
+                    }
                     if (av == null)
                     {
-                        av = new Csg.Vertex(a, new Csg.Vector2D());
+                        var avv = new Vector3(a.X, a.Y, a.Z);
+                        av = new Csg.Vertex(a, new Csg.Vector2D(e1.Dot(avv), e2.Dot(avv)));
                         vertices.Add(av);
                     }
                     if (bv == null)
                     {
-                        bv = new Csg.Vertex(b, new Csg.Vector2D());
+                        var bvv = new Vector3(b.X, b.Y, b.Z);
+                        bv = new Csg.Vertex(b, new Csg.Vector2D(e1.Dot(bvv), e2.Dot(bvv)));
                         vertices.Add(bv);
                     }
                     if (cv == null)
                     {
-                        cv = new Csg.Vertex(c, new Csg.Vector2D());
+                        var cvv = new Vector3(c.X, c.Y, c.Z);
+                        cv = new Csg.Vertex(c, new Csg.Vector2D(e1.Dot(cvv), e2.Dot(cvv)));
                         vertices.Add(cv);
                     }
 
-                    // TODO: Add texture coordinates.
                     var p = new Csg.Polygon(new List<Csg.Vertex>() { av, bv, cv });
                     polygons.Add(p);
                 }