+++ CHANGELOG.md
@@ -12,22 +12,32 @@
 - `Line.ExtendTo(IEnumerable<Line>)`
 - `Line.ExtendTo(Polyline)`
 - `Line.ExtendTo(Profile)`
+- `Line.ExtendTo(Polygon)`
 - `ConvexHull.FromPoints(IEnumerable<Vector3>)`
 - `ConvexHull.FromPolyline(Polyline)`
 - `ConvexHull.FromPolylines(IEnumerable<Polyline>)`
 - `ConvexHull.FromProfile(Profile)`
 - `Polygon.FromAlignedBoundingBox2d(IEnumerable<Vector3>)`
+- `Grid2d(Polygon, Grid1d, Grid1d)`
+- `Grid2d(IEnumerable<Polygon>, Grid1d, Grid1d)`
+- `Grid2d(Polygon, Vector3, Vector3, Vector3)`
+- `Grid2d(IEnumerable<Polygon>, Vector3, Vector3, Vector3)`
 - `Random.NextColor()` and `Random.NextMaterial()`
-
+- `Validator.DisableValidationOnConstruction`
+- `Vector3.ComputeDefaultBasisVectors()`
 ### Changed
 - Make MeshConverter deserialization more flexible to accommodate a schema used in function `input_schema`.
 - Prevent the Polygon / Polyline constructors from throwing an exception on duplicate vertices, by removing duplicates automatically.
 - Make `Grid1d` and `Grid2d` serializable
 - `new Transform(Vector3 origin, Vector3 xAxis, Vector3 yAxis, Vector3 zAxis)` did not unitize its axes, this is fixed.
+- All solids and csgs will now have planar texture coordinates.
+- Triangles are now validated to check for 3 distinct vertex positions.
+
 
 ### Fixed
 - Fixed a bug where `Polygon.UnionAll` was sometimes returning null when it shouldn't (Thanks @M-Juliani !)
 - Fixed [#517](https://github.com/hypar-io/Elements/issues/517)
+- Fixed a bug where Grid2d subcells would not split correctly with `SplitAtPoint`
 ## 0.8.2
 
 ### Changed
@@ -39,7 +49,7 @@
 
 - Fixed #483 `Deserialization of profiles created in UpdateRepresentation`
 - Fixed #484 `Failure to deserialize Model if any assembly can't be loaded.`
-- Fixed an issue where updates to a `Grid2d`'s component `Grid1d` axes would not propagate to the `Grid2d`.  
+- Fixed an issue where updates to a `Grid2d`'s component `Grid1d` axes would not propagate to the `Grid2d`.
 
 ### Added